Feature: IMR related bug fixes are resolved
	- Pregnancy and Pregnancy ANC related bug fixes are resolved
	- MDT-15362

+++ Spice-SL/app/src/main/java/com/medtroniclabs/spice/ui/medicalreview/labTechnician/NCDLabTestListActivity.kt
@@ -170,7 +170,7 @@ class NCDLabTestListActivity : BaseActivity(), View.OnClickListener, Investigati
                     "${score}%",
                     patientDetails.cvdRiskLevel, separator = getString(R.string.separator_hyphen)
                 )
-                val textColor = CommonUtils.cvdRiskColorCode(score, this)
+                val textColor = CommonUtils.cvdRiskColorCode(score.toLong(), this)
                 binding.tvCVD.setTextColor(textColor)
             }
             data.firstName?.let { firstName ->
